+++ example/src/components/InfiniteAutocomplete/index.js
@@ -1,0 +1,3 @@
+import InfiniteAutocomplete from './InfiniteAutocomplete'
+
+export default InfiniteAutocomplete