+++ example/src/components/InfiniteInput/index.js
@@ -1,0 +1,3 @@
+import InfiniteInput from './InfiniteInput'
+
+export default InfiniteInput